softdesk_support/requirements.txt
asgiref==3.7.2
certifi==2024.6.2
distlib==0.3.8
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
filelock==3.15.4
flake8==7.1.0
flake8-html==0.4.3
install==1.3.5
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
pipenv==2024.0.1
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.26.3
